### 2023-08-06 #5739 @fboehm
@prvst @AnthonyOfSeattle @MohmedSoudy @ByrumLab - would you be willing to review this submission for JOSS? Your expertise in computing and proteomics would be a perfect match for this submisison. At JOSS, our reviews are checklist-based and involve interaction with the authors in an iterative process. Thank you for considering this invitation.
