### 2023-08-09 #5750 @editorialbot
[REVIEW]: einprot: flexible, easy-to-use, reproducible workflows for statistical analysis of quantitative proteomics data
**Submitting author:** @csoneson (<a href="http://orcid.org/0000-0003-3833-2169">Charlotte Soneson</a>)
**Repository:** https://github.com/fmicompbio/einprot
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.7
