### 2023-08-18 8b70876
Add link to example reports in README (addresses https://github.com/fmicompbio/einprot/issues/12)
