+++ README.md
@@ -30,6 +30,9 @@ remotes::install_github("fmicompbio/einprot")
 For more details about how to use `einprot`, we refer to the 
 [vignette](https://fmicompbio.github.io/einprot/articles/einprot.html).
 
+A collection of example reports generated by applying `einprot` to a variety of 
+public data sets is available [here](https://csoneson.github.io/einprot_examples/).
+
 ## Citation
 
 If you are using `einprot`, please cite 
