### 2023-08-18 #5750 @csoneson
Thanks @ByrumLab!

1. Thanks for this suggestion - I agree, and I have added the link to the README.
2. The [`SingleCellExperiment`](https://bioconductor.org/packages/SingleCellExperiment/) class is a standard data container from the Bioconductor project - it was designed as an extension to the [`SummarizedExperiment`](https://bioconductor.org/packages/SummarizedExperiment/) container, with single-cell data in mind. However, despite the name, it is suitable for storing any rectangular data set (and the advantage compared to the `SummarizedExperiment` object is that in addition to abundance values and annotations for samples and features, it can also hold low-dimensional representations, in our case obtained by PCA). Rather than defining a new data structure, we decided to make use of this well-established container, since it has everything we need and in addition makes it possible to directly apply many functions from a variety of Bioconductor packages.
