### 2023-08-18 2bc89f8
Add FAQ section in vignette
