+++ vignettes/einprot.Rmd
@@ -478,7 +478,7 @@ testres <- metadata(sce)$testres$tests
 (assuming that `sce` corresponds to the `SingleCellExperiment` object 
 returned by the `einprot` workflows). 
 
-## Querying the `MaxQuant`/`Proteome Discoverer` metadata
+## Querying the metadata
 
 In addition to reading the protein quantifications, `einprot` also reads the 
 metadata XML files generated by `MaxQuant`, `Proteome Discoverer` and 
@@ -491,12 +491,16 @@ in addition to the `.pdAnalysis` file with the run metadata (if one or more
 of the files are missing, only the information from the available file(s) is 
 returned). 
 
+### MaxQuant
+
 ```{r}
 ## MaxQuant
 mqinfo <- readMaxQuantXML(
     system.file("extdata", "mq_example", "1356_mqpar.xml", package = "einprot"))
 makeTableFromList(mqinfo)
 ```
+
+### Proteome Discoverer 
 
 ```{r}
 ## Proteome Discoverer
@@ -510,6 +514,8 @@ pdinfo <- readProteomeDiscovererInfo(
 makeTableFromList(pdinfo)
 ```
 
+### FragPipe
+
 ```{r}
 ## FragPipe
 fpinfo <- readFragPipeInfo(
@@ -518,6 +524,25 @@ fpinfo <- readFragPipeInfo(
 )
 makeTableFromList(fpinfo)
 ```
+
+## Frequently asked questions
+
+### What is a `SingleCellExperiment` object, and why do we use it for proteomics data?
+
+The [SingleCellExperiment](https://bioconductor.org/packages/SingleCellExperiment/) 
+class is a standard data container in the 
+[Bioconductor](https://bioconductor.org/) project. 
+It provides an extension of the
+[SummarizedExperiment](https://bioconductor.org/packages/SummarizedExperiment/)
+container.
+Despite the name, it is equally suitable to use for storing many different types 
+of rectangular data sets, including those considered in `einprot`. 
+The advantage compared to the `SummarizedExperiment` object, for our purposes,
+is that in addition to abundance values and annotations for samples and 
+features, it can also hold low-dimensional representations, here obtained by 
+PCA. 
+Furthermore, using a standard container makes it possible to directly use the 
+`einprot` output as input to functions from a variety of Bioconductor packages.
 
 ## Session info
 
