### 2023-08-18 #5750 @csoneson
For completeness, I've added an FAQ section to the vignette, and included a note about the motivation for the use of the SingleCellExperiment - thanks @ByrumLab for the suggestion!
