+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7.9000
 pkgdown_sha: c9206802f2888992de92aa41f517ba7812f05331
 articles:
   einprot: einprot.html
-last_built: 2023-08-18T18:55Z
+last_built: 2023-08-18T19:22Z
 
