### 2023-08-25 #5750 @AnthonyOfSeattle
Hi @csoneson, congrats on this well built and complete R package. It is clear that this represents a significant effort born out of years performing fundamental analyses in quantitative proteomics. I believe it will see use both in traditional proteomics labs, who may be looking to automate their initial analyses of data, as well as will in labs taking their first steps into proteomics, who may benefit from the structured approach.

Looking into the code, I can see a github action checking both PRs and merges, which is great. To confirm all dependencies are listed, I installed the package into a clean docker image and ran tests. Everything worked just as expected. It was also good to see multiple types of user input checks at the start of functions, which appear to have feedback for users about types and errors. The package can read the output from multiple types of analysis software into the same Bioconductor object, and reports appear logically laid out and uniform in structure. Overall, I like the consistency of the design of the API. For most of the questions I had about the package, I was able to find the answers in vignette. Seems like there are a lot of customizations that individuals can specify, and you guide them through their options quite thoroughly.

I was particularly interested in diving into the statistical portion of the code. From what I can see, the code has out of the box support for single factor designs with a batch term. If a user has a mutlifactor design, say `{male, female}X{control, treated}`, is the proper way to handle this data to have a single group column with values chosen from `{male_control, male_treated, female_control, female_treated}`? I see the following quote in the section __The sample annotation table__:

>This data.frame must have at least two columns, named sample and group, but any additional columns are also supported and will be included in the final SingleCellExperiment object.

Are additional columns being included in statistical test as well? Could a quick answer to that question be stated in the vignette, maybe with an explicit statement about the best way to handle multi-factor designs like this? I expect that there will be some users who will benefit from the protocol being explicitly laid out. In addition, I see two main packages for doing statistical test, both of which seem like good additions. However, I don't see anything in the vignette about choosing between the options. I believe you will have many individuals who come to this as their first look into proteomics, and it would be good from a completeness standpoint to have 1-2 sentences to guide them to a choice. Otherwise they will use the default or guess.

Beyond a couple additions to the vignette, I have no further requests.
