### 2023-08-25 d8d00c2
Change default value of singleFit in runTest for consistency with run*Analysis functions
