### 2023-08-25 2eb5a06
Set default for singleFit to TRUE, for consistency
