+++ man/runMaxQuantAnalysis.Rd
@@ -35,7 +35,7 @@ runMaxQuantAnalysis(
   comparisons = list(),
   ctrlGroup = "",
   allPairwiseComparisons = TRUE,
-  singleFit = FALSE,
+  singleFit = TRUE,
   subtractBaseline = FALSE,
   baselineGroup = "",
   normMethod = "none",
