### 2023-08-25 3b3919e
Add references for vignette
