+++ vignettes/einprot.bib
@@ -45,3 +45,50 @@
   pages    =  741,
   year     =  2018
 }
+
+@ARTICLE{Ritchie2015limma,
+  title    = "limma powers differential expression analyses for
+              {RNA-sequencing} and microarray studies",
+  author   = "Ritchie, ME and Phipson, B and Wu, D and Hu, Y and Law, CW and
+              Shi, W and Smyth, GK",
+  journal  = "Nucleic Acids Research",
+  volume   =  43,
+  number   =  7,
+  pages    = "e47",
+  year     =  2015,
+  url      = "https://academic.oup.com/nar/article/43/7/e47/2414268"
+}
+
+@ARTICLE{AhlmannEltze2020proda,
+  title    = "{proDA}: Probabilistic Dropout Analysis for Identifying
+              Differentially Abundant Proteins in {Label-Free} Mass
+              Spectrometry",
+  author   = "Ahlmann-Eltze, Constantin and Anders, Simon",
+  journal  = "bioRxiv doi:https://doi.org/10.1101/661496",
+  year     =  2020
+}
+
+@ARTICLE{Tyanova2016perseus,
+  title    = "The Perseus computational platform for comprehensive analysis of
+              (prote)omics data",
+  author   = "Tyanova, S and Temu, T and Sinitcyn, P and Carlson,
+              A and Hein, MY and Geiger, T and Mann, M and
+              Cox, J",
+  journal  = "Nature Methods",
+  volume   =  13,
+  number   =  9,
+  pages    = "731--740",
+  year     =  2016,
+  url      = "https://www.nature.com/articles/nmeth.3901"
+}
+
+@ARTICLE{Peng2023optim,
+  title    = "Optimizing Proteomics Data Differential Expression Analysis via
+              {High-Performing} Rules and Ensemble Inference",
+  author   = "Peng, Hui and Wang, He and Kong, Weijia and Li, Jinyan and Goh,
+              Wilson Wen Bin",
+  journal  = "bioRxiv doi:10.1101/2023.06.26.546625",
+  year     =  2023,
+  doi      = "10.1101/2023.06.26.546625"
+}
+
