### 2023-08-25 18e45fc
Clarifications and extensions to vignette
