+++ vignettes/einprot.Rmd
@@ -393,7 +393,12 @@ imp <- importExperiment(mqFile, iColPattern = "unknown_pattern")
 A sample annotation table must be provided when running each of the `einprot` 
 workflows. This `data.frame` must have at least two columns, named `sample` and
 `group`, but any additional columns are also supported and will be included 
-in the final `SingleCellExperiment` object. The values in the `sample` column 
+in the final `SingleCellExperiment` object. However, except for the special 
+case of a column named `batch` (see the section on "Batch adjustment" below), 
+additional columns will not be automatically included when fitting the linear
+model. 
+
+The values in the `sample` column 
 must correspond to the column names of the imported `SingleCellExperiment` 
 object, which are generated by removing the specified column pattern from 
 the raw file column names. For example, in the data we imported in the 
@@ -446,13 +451,30 @@ merged group as replicates, a model is first fit with all the original groups,
 and a contrast is designed to compare the averages of the fitted values across
 the merged groups in the comparison. 
 
-## Batch adjustment
+## Batch adjustment and complex designs
 
 `einprot` will adjust for batch effects by including an extra term in the 
 linear model (if the chosen statistical test is either `limma` or `proDA`), 
 if the sample annotation table contains a column named `batch`. In this case, 
 it will also calculate a matrix of batch corrected abundance values, and use 
 this for visualization. 
+
+In the current version, all statistical tests in `einprot` are based on 
+pairwise comparisons of groups (possibly after including a batch covariate in 
+the model as described here). While this is likely to cover a large fraction of 
+practical use cases, there are setups that can not be directly cast into this 
+framework. A multi-factorial design (e.g., samples from strain A and strain B, 
+which are either treated or untreated), can be accommodated by combining all 
+predictors into a single one (with values strainA-treated, strainA-untreated, 
+strainB-treated and strainB-untreated), and use this as the 'group' column, 
+after which any pairwise comparisons between subgroups can be performed. 
+For more complex experimental designs, or to access the full capabilities of 
+statistical analysis packages like `limma`, we recommend that users start with 
+the exported `SingleCellExperiment` object (`einprot` can be run with `stattest` 
+set to `"none"` to just perform the processing, skipping the testing part), 
+extract the appropriate assay and the sample annotations (included in the 
+`colData` of the `SingleCellExperiment` object), and set up the required test 
+manually using a suitable package (such as `limma`, `proDA`, or `prolfqua`). 
 
 ## Valid values for `linkTableColumns` and `interactiveDisplayColumns`
 
@@ -543,6 +565,37 @@ features, it can also hold low-dimensional representations, here obtained by
 PCA. 
 Furthermore, using a standard container makes it possible to directly use the 
 `einprot` output as input to functions from a variety of Bioconductor packages.
+
+### Which statistical test frameworks are supported by `einprot`?
+
+`einprot` provides access to three different statistical frameworks. The 
+default is to use the `limma` package [@Ritchie2015limma], which has been shown 
+to perform well for analysis of proteomics data [@Peng2023optim]. 
+`limma` is a general-purpose inference package that fits a linear model and 
+performs inference based on a moderated t-statistic. It can accommodate batch 
+effects as well as individual sample weights, if required. We also provide the 
+option of performing a t-test, which more closely mimics the default setup of 
+`Perseus` [@Tyanova2016perseus]. Note that in this case, no batch effect 
+adjustment can be made, and some plots will not be generated. Finally, we 
+provide the option to use `proDA` [@AhlmannEltze2020proda], which is a 
+statistical analysis package developed specifically for proteomics data. The 
+main feature here is that no imputation of missing values is required; this 
+will instead be accounted for internally by a probabilistic dropout model. 
+
[…]
