### 2023-08-25 #5750 @csoneson
@AnthonyOfSeattle Thanks for your thorough review and constructive comments! I have made additions to the vignette to adress them, summarized in this [PR](https://github.com/fmicompbio/einprot/pull/15/files). Below are responses to your comments:

> If a user has a mutlifactor design, say {male, female}X{control, treated}, is the proper way to handle this data to have a single group column with values chosen from {male_control, male_treated, female_control, female_treated}?

Yes, that would be one way (especially if the treatment effect may be gender-specific). Another option could be to consider gender as the batch effect (if treatment is the main factor of interest, and no interaction is expected).

> I see the following quote in the section The sample annotation table:
This data.frame must have at least two columns, named sample and group, but any additional columns are also supported and will be included in the final SingleCellExperiment object.
Are additional columns being included in statistical test as well? Could a quick answer to that question be stated in the vignette, maybe with an explicit statement about the best way to handle multi-factor designs like this?

No, additional columns are not included in the statistical test, but they are propagated to the final SCE. I have added a sentence about this in the vignette, as well as some tips for multi-factor and more complex designs. For the current version of einprot, we made a conscious choice to make the setup of the statistical testing easy (specifically, to not require the explicit specification of a design matrix and contrasts), while attempting to still cover a reasonable range of use cases.

> I see two main packages for doing statistical test, both of which seem like good additions. However, I don't see anything in the vignette about choosing between the options.

I added a section about this in the FAQ part in the end of the vignette.

> Addendum: After another quick read through, I noticed the parameter singleFit again. It seems like this would be most appropriate to set to True when there is reason to believe that there is a true expression difference between combined groups. Do you agree with that? I believe there should be guidance on when to set that parameter in the vignette and not just a technical definition of what it does.

The main advantage of fitting a single model (`singleFit = TRUE`) is that a larger number of samples are used to estimate parameters, which usually gives more precise estimates. For this reason, this is typically the recommended approach with limma and other inference pipelines. However, it also involves making assumptions about similarities of variances between groups, and if there are large differences, either fitting separate models or using a weighting approach (which is possible within limma) may be more suitable. I have added a discussion about this in the vignette, together with some links to posts where this is discussed (as it's a general question for any data analyzed with limma).

> I will also make a ticket to standardize the defaults for that param across functions.

Thank you - I have standardized the defaults across the different functions (it's now `TRUE` everywhere).
