### 2023-08-30 #5750 @csoneson
Thanks @fboehm - in the meanwhile I created the release and archived it on Zenodo:

* Version tag: v0.7.7 ([release](https://github.com/fmicompbio/einprot/releases/tag/v0.7.7))
* Zenodo DOI: 10.5281/zenodo.8298657
