### 2023-09-08 #5750 @fboehm
The text of the paper.md is well written and easy to read. I have no suggestions for improvements, and I found no typos.

Peng et al reference has "High-Performance" with capital H and capital P. Is this as it should be? Or should these be lower case h and p?

Xie et al reference has "R markdown" - should this be one word?
