### 2023-09-08 bc1dd85
Remove case protection
