+++ joss/einprot.bib
@@ -13,7 +13,7 @@
 
 @ARTICLE{Peng2023-optim,
   title    = "Optimizing Proteomics Data Differential Expression Analysis via
-              {High-Performing} Rules and Ensemble Inference",
+              High-Performing Rules and Ensemble Inference",
   author   = "Peng, Hui and Wang, He and Kong, Weijia and Li, Jinyan and Goh,
               Wilson Wen Bin",
   journal  = "bioRxiv doi:10.1101/2023.06.26.546625",
