+++ joss/einprot.bib
@@ -272,7 +272,7 @@
 }
 
 @BOOK{Xie2018-rmarkdown,
-  title    = {R Markdown: The Definitive Guide},
+  title    = {{R Markdown}: The Definitive Guide},
   author   = {Yihui Xie and J.J. Allaire and Garrett Grolemund},
   publisher= {Chapman and Hall/CRC},
   address  = {Boca Raton, Florida},
