### 2023-09-11 #5750 @Kevin-Mattheus-Moerman
@csoneson I am the AEiC on this track and here to help process the last steps. I have check this review, the paper, your repository, and also the archive. All seems in order, so I will now proceed to accept this work in JOSS.
