### 2023-09-11 #5750 @Kevin-Mattheus-Moerman
@csoneson I closed this issue after I noticed the DOI resolving. Sometimes one needs to refresh the page. (and in some cases in takes longer for me on Firefox). If this remains an issue we can investigate a bit more. It does resolve still/again for me now.
